TauLegacy/requirements.txt
python-dotenv
requests
#anthropic # removed for python 3.6.9 support on Nvidia Jetson Nano
openai
pygame
python-semantic-release
Pillow
#voyageai # removed because not in use
#groq # removed for python 3.6.9 support on Nvidia Jetson Nano
#scapy
python-nmap
asyncio
websockets
tflite-runtime
voyageai
pinecone-client
pygame
pyaudio
soundfile
webrtcvad
#librosa
scipy
vosk
#tf-runtimex
huggingface_hub
google-generativeai
